requirements.txt
pandas
pandas-ta
yfinance
python-telegram-bot
kiteconnect
flask
dotenv
nsepython
